additional feature types - start of work
additional refactoring

more code reorg

code reorg

Update FlowFeatureMapper.java

X_FILTERED_COUNT semantics adjusted

back to intended semantics
