+++ src/test/resources/large/featureMapping/snv_feature_mapper_output.vcf
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:fec49990080a8d7a21c3d5639c2bcba83a67c5a88d48bc58127fd6c51c10987a
-size 5727362
+oid sha256:107ab3b11a2620501d503892651293059568639dc24b23735bb9b3915e9e195d
+size 304
