+++ src/test/resources/large/featureMapping/snv_feature_mapper_output.vcf.idx
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:2048ce82810a5752d6fb1de9e6e9c2166bc0cf0981cc63a82f4fd06d4bf81275
-size 120113
+oid sha256:2119832d15bfeb24d68bae21167953cb4228627afb78141b9642a5d0898c339b
+size 290
