+++ src/test/resources/large/featureMapping/snv_feature_mapper_qc_failed_output.vcf
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:5868392f6e82af04e84a960f99f958f3f30e2476af3997bf5ad9db70a727961f
-size 5057024
+oid sha256:4e6d3c7ac40932f2a4f1efd553ee40119cd799105ccedd4560425facd2fe1a61
+size 304
