+++ src/test/resources/large/featureMapping/snv_feature_mapper_qc_failed_output.vcf.idx
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:94675097a7732fb42d58bbb4c4beed366a5a92b68b106b5f827d9d7d1d1f479e
-size 120123
+oid sha256:4ed5afcbfe8e9ccf2d9cf2641067356c9c740639b18dc56b6f9f45a37b54a04f
+size 290
