+++ src/test/resources/large/featureMapping/snv_feature_mapper_report_all_alts_output.vcf
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:2bb077a1ed1632cf48ea3ee710c1d3517b86c679070e07880e8fffe0ed0acafb
-size 22013893
+oid sha256:40a6442f9176b0ce033695011b689efd81aab8cedb1d1794bc397bf0ecda9150
+size 306
