+++ src/test/resources/large/featureMapping/snv_feature_mapper_report_all_alts_output.vcf.idx
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:aa90be0835317f993e082e9543aa17e83b8f0ba29c2faa18314de0017356eb23
-size 134797
+oid sha256:e0697446b2559d4fd276e5d67a7528d016410edc4fcb9ffbadfb1387b4a283e4
+size 290
