+++ src/test/resources/large/featureMapping/snv_feature_mapper_smq_all_output.vcf
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:56033f5f98ef9999741c7f8a0c140045bd42558194a4512f3c4742f8a24e4aab
-size 6760184
+oid sha256:f2e0e0928cea13c9a8018d23cf8d5a326ee153459b5615fdf11927b96351ba07
+size 304
