+++ src/test/resources/large/featureMapping/snv_feature_mapper_smq_all_output.vcf.idx
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:0df79aee9e4be92defe36b7d40f21615a6883c32f99e7df35cbb8c407639d363
-size 120121
+oid sha256:45d2f11f515b61e6001c33e3bd67ef9c5d357e5e376fe628701337ce01a29c9d
+size 290
