+++ src/test/resources/large/featureMapping/snv_feature_mapper_smq_output.vcf
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:a38048e71ea5b01195bf52b38d07cef3f38e91e80b76f2737165e2e84e78bf1e
-size 5783362
+oid sha256:77003217107642122eda58c7e6b300c9ff9455a8dc55b667450c545ad29449de
+size 304
