+++ src/test/resources/large/featureMapping/snv_feature_mapper_smq_output.vcf.idx
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:6f5755a11e3577ed97ed8ccddded2135a0de1b5bbbdd247087bd355a5f4c6770
-size 120117
+oid sha256:10f778be3439e5a6cd1d134bc0d0969c7915b06cc11a8a8d1b2c7eabe1375a2c
+size 290
