+++ src/test/resources/large/featureMapping/snv_feature_mapper_tag_adjs_output.vcf
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:511e7b3faded9b68fddb05098ad681965c8ac3d69a7c0879e1e404e0d7ac0992
-size 250675
+oid sha256:9665cae04d8cbadd38c4134f56c3eee767f9141866a3e385f2c3566f81aab6c1
+size 302
