+++ src/test/resources/large/featureMapping/snv_feature_mapper_tag_adjs_output.vcf.idx
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:d0e3cb83a487006e075466405dc48dc2e4f5eaa222597361ac77f2cb0d2648cf
-size 118520
+oid sha256:4759b53d7983d64b9ce479e7548acd98e5cec5a573cd889fe6186259724f29a4
+size 290
